Revert "All bugs are fixed"
This reverts commit 5440f7ec32b65c11f5fd189085bf6f27a459bf01.
