+++ client/src/context/actions.js
@@ -72,6 +72,8 @@ export async function logout(dispatch) {
   localStorage.removeItem('profile_pic');
   localStorage.removeItem('list_of_products');
   localStorage.removeItem('friends');
+  localStorage.removeItem('randomUsers');
+  localStorage.removeItem('friendDetails');
 
 }
 
@@ -146,6 +148,54 @@ export async function unfollowFriends(dispatch, token, friend) {
   fetchData();
 
 }
+
+export async function getRandomUsers(myfriends) {
+  async function fetchData() {
+    const input = {
+      getFriends : false,
+      friends : myfriends
+    }
+
+    const response = await fetch(`/users`, {
+      method: "PUT",
+      headers: {
+        "Content-Type": "aplication/json",
+      },
+      body: JSON.stringify(input),
+    });
+    const randomUsers = await response.json();
+
+    localStorage.setItem("randomUsers", JSON.stringify(randomUsers))
+
+  };
+  fetchData();
+
+}
+
+export async function getFriendDetails(myfriends) {
+  async function fetchData() {
+    const input = {
+      getFriends : true,
+      friends : myfriends
+    }
+
+    const response = await fetch(`/users`, {
+      method: "PUT",
+      headers: {
+        "Content-Type": "aplication/json",
+      },
+      body: JSON.stringify(input),
+    });
+    const friendDetails = await response.json();
+    console.log(friendDetails);
+
+    localStorage.setItem("friendDetails", JSON.stringify(friendDetails))
+
+  };
+  fetchData();
+
+}
+
 
 
 /*
